[1] validation results now explicitly casted to geodataframe so that to_file works for tabular datasets; [2] validation `dates_range` casts to int64 when suffixing dates with 01 since they may be stored as incompatible int32 from input

+++ src/validate/validation_functions.py
@@ -469,7 +469,10 @@ class Validator:
         # Fetch current date.
         today = int(datetime.today().strftime("%Y%m%d"))
 
-        # Temporary populate incomplete dates with "01" suffix.
+        # Temporarily populate incomplete dates with "01" suffix.
+        # Note: Requires casting to int64.
+        series = series.astype("int64")
+
         for length in (4, 6):
             flag = series.map(lambda val: int(math.log10(val)) + 1) == length
             if length == 4:
